@@ -4,26 +4,27 @@
     "license": "MIT",
     "require": {
         "php": "^8.1",
+        "ext-mbstring": "*",
+        "ext-tokenizer": "*",
         "composer/xdebug-handler": "^3.0",
         "symfony/yaml": "^5.1",
         "phpactor/container": "^3.0",
         "phpactor/class-to-file": "~0.5",
         "twig/twig": "^3.4",
         "dnoegel/php-xdg-base-dir": "^0.1.0",
-        "symfony/console": "^5.0",
+        "symfony/console": "^6.0",
         "dantleech/invoke": "^2.0",
         "phpactor/amp-fswatch": "^0.3.0",
         "amphp/process": "^1.1.5",
-        "thecodingmachine/safe": "^1.0",
-        "phpactor/phly-event-dispatcher": "^2.0.0",
-        "phpactor/language-server": "dev-master",
-        "phpactor/language-server-protocol": "^3.17.3",
+        "phpactor/phly-event-dispatcher": "^2.2.0",
+        "phpactor/language-server": "^7.0.1",
+        "phpactor/language-server-protocol": "^3.17.4",
         "dantleech/object-renderer": "^0.1.1",
         "monolog/monolog": "^1.23",
         "sebastian/diff": "^4.0",
         "webmozart/glob": "^4.4",
-        "symfony/filesystem": "^5.4",
-        "symfony/process": "^5.4",
+        "symfony/filesystem": "^6.0",
+        "symfony/process": "^6.0",
         "jetbrains/phpstorm-stubs": "dev-master",
         "phpactor/tolerant-php-parser": "dev-main",
         "phpactor/map-resolver": "^1.5.0",
@@ -37,13 +38,13 @@
         "phpspec/prophecy-phpunit": "^2.0",
         "phpactor/test-utils": "^1.1.4",
         "symfony/var-dumper": "^6.4",
-        "phpbench/phpbench": "^1.1",
+        "phpbench/phpbench": "^1.4",
         "dms/phpunit-arraysubset-asserts": "dev-master",
         "psr/log": "^1.1",
         "phpstan/phpstan-phpunit": "^1.0",
         "jangregor/phpstan-prophecy": "^1.0",
         "phpstan/extension-installer": "^1.1",
-        "psalm/phar": "^5.15",
+        "psalm/phar": "^6.0",
         "blackfire/php-sdk": "^1.31",
         "squizlabs/php_codesniffer": "^3.7",
         "rector/rector": "^1.2"
