@@ -4,6 +4,7 @@ namespace Phpactor\Extension\LanguageServerCodeTransform;
 
 use Phpactor\ClassFileConverter\Domain\FileToClass;
 use Phpactor\CodeTransform\Adapter\WorseReflection\Refactor\WorseFillMatchArms;
+use Phpactor\CodeTransform\Adapter\TolerantParser\Refactor\TolerantHereDoc;
 use Phpactor\CodeTransform\Adapter\WorseReflection\Refactor\WorseFillObject;
 use Phpactor\CodeTransform\Domain\Generators;
 use Phpactor\CodeTransform\Domain\Helper\MissingMemberFinder;
@@ -266,7 +267,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 $container->get(ReplaceQualifierWithImport::class)
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [],
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100],
         ]);
 
         $container->register(ImportNameProvider::class, function (Container $container) {
@@ -275,7 +276,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 $container->getParameter(self::PARAM_REPORT_NON_EXISTING_NAMES)
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [],
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100],
         ]);
         $container->register(OverridableMethodFinder::class, function (Container $container) {
             return new OverridableMethodFinder(
@@ -287,7 +288,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 $container->get(OverridableMethodFinder::class),
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [],
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => ['priority' => -200],
         ]);
 
         $container->register(TransformerCodeActionPovider::class.'promote_constructor_private', function (Container $container) {
@@ -297,7 +298,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'Promote Constructor (private)'
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
 
         $container->register(TransformerCodeActionPovider::class.'promote_constructor_public', function (Container $container) {
@@ -307,7 +308,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'Promote Constructor (public)'
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
 
         $container->register(TransformerCodeActionPovider::class.'complete_constructor_private', function (Container $container) {
@@ -317,7 +318,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'Complete Constructor (private)'
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
 
         $container->register(TransformerCodeActionPovider::class.'complete_constructor_public', function (Container $container) {
@@ -327,7 +328,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'Complete Constructor (public)'
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
         $container->register(TransformerCodeActionPovider::class.'add_missing_class_generic', function (Container $container) {
             return new TransformerCodeActionPovider(
@@ -434,7 +435,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'Remove unused imports'
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
 
         $container->register(GenerateMemberProvider::class, function (Container $container) {
@@ -498,7 +499,7 @@ class LanguageServerCodeTransformExtension implements Extension
                 'fill new object construct with named parameters',
             );
         }, [
-            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
         $container->register(ByteOffsetRefactorProvider::class.'.fill_match_arms', function (Container $container) {
             return new ByteOffsetRefactorProvider(
@@ -509,6 +510,16 @@ class LanguageServerCodeTransformExtension implements Extension
             );
         }, [
             LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => []
+        ]);
+        $container->register(ByteOffsetRefactorProvider::class.'.here_doc_provider', function (Container $container) {
+            return new ByteOffsetRefactorProvider(
+                $container->get(TolerantHereDoc::class),
+                'quickfix.here_doc_provider',
+                'Convert HereDoc',
+                'replace string with HereDoc',
+            );
+        }, [
+            LanguageServerExtension::TAG_CODE_ACTION_PROVIDER => [ 'priority' => 100]
         ]);
 
         $container->register(GenerateConstructorProvider::class, function (Container $container) {
