@@ -2,17 +2,14 @@
 
 namespace Phpactor\Extension\LanguageServerReferenceFinder;
 
-use Microsoft\PhpParser\Parser;
 use Phpactor\Container\Container;
 use Phpactor\Container\ContainerBuilder;
 use Phpactor\Container\Extension;
 use Phpactor\Extension\LanguageServerBridge\Converter\LocationConverter;
 use Phpactor\Extension\LanguageServerReferenceFinder\Handler\GotoDefinitionHandler;
 use Phpactor\Extension\LanguageServerReferenceFinder\Handler\GotoImplementationHandler;
-use Phpactor\Extension\LanguageServerReferenceFinder\Handler\HighlightHandler;
 use Phpactor\Extension\LanguageServerReferenceFinder\Handler\ReferencesHandler;
 use Phpactor\Extension\LanguageServerReferenceFinder\Handler\TypeDefinitionHandler;
-use Phpactor\Extension\LanguageServerReferenceFinder\Model\Highlighter;
 use Phpactor\Extension\LanguageServerReferenceFinder\Adapter\Indexer\WorkspaceUpdateReferenceFinder;
 use Phpactor\Extension\LanguageServer\LanguageServerExtension;
 use Phpactor\Extension\ReferenceFinder\ReferenceFinderExtension;
@@ -24,6 +21,7 @@ use Phpactor\ReferenceFinder\ReferenceFinder;
 class LanguageServerReferenceFinderExtension implements Extension
 {
     const PARAM_REFERENCE_TIMEOUT = 'language_server_reference_reference_finder.reference_timeout';
+    const PARAM_REFERENCE_SOFT_TIMEOUT = 'language_server_reference_finder.soft_timeout';
 
 
     public function load(ContainerBuilder $container): void
@@ -61,7 +59,8 @@ class LanguageServerReferenceFinderExtension implements Extension
                 $container->get(ReferenceFinderExtension::SERVICE_DEFINITION_LOCATOR),
                 $container->get(LocationConverter::class),
                 $container->get(ClientApi::class),
-                $container->getParameter(self::PARAM_REFERENCE_TIMEOUT)
+                $container->parameter(self::PARAM_REFERENCE_TIMEOUT)->int(),
+                $container->parameter(self::PARAM_REFERENCE_SOFT_TIMEOUT)->int(),
             );
         }, [ LanguageServerExtension::TAG_METHOD_HANDLER => [] ]);
 
@@ -72,23 +71,18 @@ class LanguageServerReferenceFinderExtension implements Extension
                 $container->get(LocationConverter::class)
             );
         }, [ LanguageServerExtension::TAG_METHOD_HANDLER => [] ]);
-
-        $container->register(HighlightHandler::class, function (Container $container) {
-            return new HighlightHandler(
-                $container->get(LanguageServerExtension::SERVICE_SESSION_WORKSPACE),
-                new Highlighter(new Parser())
-            );
-        }, [ LanguageServerExtension::TAG_METHOD_HANDLER => [] ]);
     }
 
 
     public function configure(Resolver $schema): void
     {
         $schema->setDefaults([
-            self::PARAM_REFERENCE_TIMEOUT => 60
+            self::PARAM_REFERENCE_TIMEOUT => 60,
+            self::PARAM_REFERENCE_SOFT_TIMEOUT => 10,
         ]);
         $schema->setDescriptions([
             self::PARAM_REFERENCE_TIMEOUT => 'Stop searching for references after this time (in seconds) has expired',
+            self::PARAM_REFERENCE_SOFT_TIMEOUT => 'Interupt and ask for confirmation to continue after this timeout (in seconds)',
         ]);
     }
 }
