@@ -3,14 +3,13 @@
 namespace Phpactor\Indexer\Tests\Adapter;
 
 use Phpactor\Indexer\Tests\IntegrationTestCase;
-use function Safe\file_get_contents;
 
 abstract class IndexTestCase extends IntegrationTestCase
 {
     protected function setUp(): void
     {
         $this->workspace()->reset();
-        $this->workspace()->loadManifest(file_get_contents(__DIR__ . '/Manifest/buildIndex.php.test'));
+        $this->workspace()->loadManifest((string)file_get_contents(__DIR__ . '/Manifest/buildIndex.php.test'));
     }
 
     public function testBuild(): void
