@@ -11,7 +11,9 @@ interface ReferenceFinder
     /**
      * Find references to the symbol at the given byte offset.
      *
-     * @return Generator<PotentialLocation>
+     * Return true if no more finders should be executed after this finder and false otherwise.
+     *
+     * @return Generator<int, PotentialLocation, null, bool>
      */
     public function findReferences(TextDocument $document, ByteOffset $byteOffset): Generator;
 }
