@@ -8,7 +8,6 @@ use Phpactor\ReferenceFinder\TypeLocation;
 use Phpactor\ReferenceFinder\TypeLocations;
 use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\Location;
-use Phpactor\TextDocument\TextDocumentUri;
 use Phpactor\WorseReflection\Core\Cache;
 use Phpactor\WorseReflection\Core\ClassHierarchyResolver;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
@@ -97,9 +96,9 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
             throw new CouldNotLocateDefinition($e->getMessage(), 0, $e);
         }
 
-        $path = $class->sourceCode()->uri();
+        $uri = $class->sourceCode()->uri();
 
-        if (null === $path) {
+        if (null === $uri) {
             throw new CouldNotLocateDefinition(sprintf(
                 'The source code for class "%s" has no path associated with it.',
                 $class->name()
@@ -107,7 +106,7 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
         }
 
         return new TypeLocations([new TypeLocation($className, new Location(
-            TextDocumentUri::fromString($path),
+            $uri,
             $class->position()
         ))]);
     }
@@ -122,9 +121,9 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
             throw new CouldNotLocateDefinition($e->getMessage(), 0, $e);
         }
 
-        $path = $function->sourceCode()->uri()?->path();
+        $uri = $function->sourceCode()->uri();
 
-        if (null === $path) {
+        if (null === $uri) {
             throw new CouldNotLocateDefinition(sprintf(
                 'The source code for function "%s" has no path associated with it.',
                 $function->name()
@@ -133,7 +132,7 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
 
         return new TypeLocations([
             new TypeLocation(TypeFactory::unknown(), new Location(
-                TextDocumentUri::fromString($path),
+                $uri,
                 $function->position()
             ))
         ]);
@@ -149,9 +148,9 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
             throw new CouldNotLocateDefinition($e->getMessage(), 0, $e);
         }
 
-        $path = $constant->sourceCode()->uri()?->path();
+        $uri = $constant->sourceCode()->uri();
 
-        if (null === $path) {
+        if (null === $uri) {
             throw new CouldNotLocateDefinition(sprintf(
                 'The source code for constant "%s" has no path associated with it.',
                 $constant->name()
@@ -160,7 +159,7 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
 
         return new TypeLocations([
             new TypeLocation(TypeFactory::unknown(), new Location(
-                TextDocumentUri::fromString($path),
+                $uri,
                 $constant->position()
             ))
         ]);
@@ -220,9 +219,9 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
 
             $member = $members->get($symbolName);
 
-            $path = $member->declaringClass()->sourceCode()->uri()?->path();
+            $uri = $member->declaringClass()->sourceCode()->uri();
 
-            if (null === $path) {
+            if (null === $uri) {
                 throw new CouldNotLocateDefinition(sprintf(
                     'The source code for class "%s" has no path associated with it.',
                     (string) $containingClass->name()
@@ -231,7 +230,7 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
 
             $locations[] = new TypeLocation(
                 $namedType,
-                new Location(TextDocumentUri::fromString($path), $member->position())
+                new Location($uri, $member->position())
             );
         }
 
@@ -261,18 +260,18 @@ class WorseReflectionDefinitionLocator implements DefinitionLocator
         }
 
 
-        $path = $member->declaringClass()->sourceCode()->uri()?->path();
+        $uri = $member->declaringClass()->sourceCode()->uri();
 
-        if (null === $path) {
+        if (null === $uri) {
             throw new CouldNotLocateDefinition(sprintf(
                 'The source code for class "%s" has no path associated with it.',
                 (string) $member->declaringClass()->name()
             ));
         }
 
-        return new TypeLocations([new TypeLocation($nodeContext->classType(), new Location(
-            TextDocumentUri::fromString($path),
-            $member->position()
-        ))]);
+        return new TypeLocations([new TypeLocation(
+            $nodeContext->classType(),
+            new Location($uri, $member->position())
+        )]);
     }
 }
